fix: Coverage report
Part of #31

Split CMakeLists.txt to 3 files, one on root, one in src and one in
tests. Each are for its own goal. It make easier to do code coverage.

In local testing code coverage still take only test file, but need to
see in future if with complete tests src files will be included or not.