feat: Give up on cross-compilation
Part of #2

Cross-compilation is a very hard topic for both for bazel and clang with
too many constraints. Give up on it for moment, will see later if a
solution is possible

So filc will be built only for x86_64 machines, if you run on another
architecture you'll need to clone the repository and built it on your
own.