refactor: Library dependencies provided by nix
Part of #54

Instead of using FetchContent in cmake to download libraries, use
nix-shell to add them as dependencies. Cmake can then retrieve them.

When in nix-shell, nix store paths are already added to cmake vars. But
when using an IDE not aware of nix environment, we need to add it
manually -> line 4 to 9 in CMakeLists.txt are here for that